+++ web/skins/modern/user/cat_info.php
@@ -97,7 +97,7 @@ switch ($page) {
                     $out .= '<script type="text/javascript">goAdmin()</script>';
                 } else {
                     if (\config\ConfAssistant::CONSORTIUM['selfservice_registration'] === NULL) {
-                        $out .= sprintf(_("You must have received an invitation from your %s %s before being able to manage your %s. If that is the case, please continue and log in."), \config\ConfAssistant::CONSORTIUM['display_name'], core\common\Entity::$nomenclature_fed, core\common\Entity::$nomenclature_inst);
+                        $out .= sprintf(_("You must have received an invitation from your %s %s before being able to manage your %s. If that is the case, please continue and log in."), \config\ConfAssistant::CONSORTIUM['display_name'], core\common\Entity::$nomenclature_fed, core\common\Entity::$nomenclature_participant);
                     } else {
                         $out .= _("Please authenticate yourself and login");
                     }